requirements.txt
Flask==2.3.2
PyPDF2==3.0.1
sentence-transformers==2.2.2
openai==0.28.0
langchain==0.14.0
chroma-hnswlib==0.14.0
python-dotenv==1.0.0
